requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg==3.2.10
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.2.0
